[tb/fix-label-image-popover] Move label image editor to its folder
